separated confirm to concern
