+++ .rubocop.yml
@@ -24,6 +24,8 @@ Lint/AmbiguousOperatorPrecedence: # new in 1.21
   Enabled: true
 Lint/AmbiguousRange: # new in 1.19
   Enabled: true
+Lint/ConstantDefinitionInBlock:
+  Enabled: false
 Lint/ConstantOverwrittenInRescue: # new in 1.31
   Enabled: true
 Lint/DeprecatedConstants: # new in 1.8
