+++ app/models/user.rb
@@ -2,8 +2,8 @@
 
 class User < ApplicationRecord
   include Rememberable
+  include Confirmable
 
-  CONFIRMATION_TOKEN_EXPIRATION = 10.minutes
   MAILER_FROM_EMAIL = 'no-reply@example.com'
 
   has_secure_password
@@ -23,27 +23,6 @@ class User < ApplicationRecord
 
   def full_name
     "#{first_name} #{last_name}"
-  end
-
-  def confirm!
-    update_columns(confirmed_at: Time.current)
-  end
-
-  def confirmed?
-    confirmed_at.present?
-  end
-
-  def generate_confirmation_token
-    signed_id expires_in: CONFIRMATION_TOKEN_EXPIRATION, purpose: :confirm_email
-  end
-
-  def unconfirmed?
-    !confirmed?
-  end
-
-  def send_confirmation_email!
-    confirmation_token = generate_confirmation_token
-    UserMailer.confirmation(self, confirmation_token).deliver_now
   end
 
   def role?(role)
